started adding a second icl7660
after testing the prototypes it was found that the voltage drop is too much, the output voltage is around 8.5v which actually works in a tr707, but it could be better.
acording to the intersill datasheet, you can just add another icl7660 in parallel to reduce voltage drop, lets try it.

also changed to the ceramic caps from 0402 to 1203

updated the schematic and pcb