Big rework.
re-added a m7 diode with a jumper to drop the voltage if required (incase the other improvements work TOO well)

increase value of charge pump capacitors to increase output voltage.

added a seccond

bigger output capacitor removes a bit more ripple. (a footprint for a THT capacitor is also provided)

respin PCB

to accomodate all the above the pcb size is increased.